Major performance improvements
Performance improvements by adding a new unordered map on secure
addresses so server requests aren't needed to made any longer, the
client still makes a single one upon join attempt as it doesn't receive
any data from the server...

The only way to solve this is to store the data in the broadcast replies
which is inefficient when there's multiple severs/lobbies running it
will allocate tons of data.

We now only populate Join request and connect-request packets with
XNADDR data in order to improve performance there.

All for loops were checked that loop the user array and changed to only
loop from 1-15 out of 16 [0] contains local user data.

It's still causing FPS drops once there's multiple users in I think it's
related to the memcpy of the buffer trying to find a way to resolve this
the only thing I can think of is attempting to allocate an additional 4
bytes to the end of each packet, this adds it's own complicates the only
way to avoid this may be to disable these checks all together and just
use XNADDR for now on it needs further research.