Complete Network Re-write - Improved Frames, Removed Memory Leaks, Currently has Network latency due to added packet.
All of the network functionality has been completely re-written from
scratch, it currently lacks NAT support and the way the
Join-request/Sending of Secure addresses is handled is currently flawed.

We need to establish a map of hosts which have already been connected to
and reference this map as if it were a boolean array, if we've never
connected to x host on y port then we send our secure addr otherwise
don't and just continue through normal packet operations.

Right now join request/connect-request are detected but  this only
handles port 1001 we need to handle 1000 and this is the only way I can
think of doing so.