Hacked together support for VS 2015, removed OSHGUI partially
So since I've moved development to my laptop I needed to get the thing
to compile, I've modified a few keypoints of the minhook in order to get
it compiling in vs 2015 and then proceeded by removing oshgui.

As long as the Direct X June SDK is installed this should compile fine
for just about anyone at this point, libprotonbuf is included in this
commit.