Disable Hitmarker
This should disable the functionality for the hitmarker effect and remove the option from the other settings menu. My opinion is that this should be disabled since it has an issue with it playing the sound on the wrong player due to player datum stuff. It has been like this since it's inception and I'm not sure if KC has any time at the moment to fix it. Best left disabled until it can be completed.