Faster Live server list downloading + misc changes
-Force Refresh Rate
Saves and forces a selected refresh rate from the Carto settings menu on startup
-Save mouse/controller sens
Saves the value last inputed into the console and pokes it on map load. The default value is zero and the code does nothing if it is zero.
-Fix for console spawn commands
Should only be able to be used by the host now
-Hitfix reverted to 4.0.4 values
-Live server list downloads faster
Original commit by Nuke found here
https://github.com/NukeULater/cartographer/commit/1d14f0d21a240ddc15517e2ab37ec171ceef017f
-Fix for Xinput copying
This fixes splitscreen. Original commit by num005 found here
https://github.com/num0005/cartographer/commit/2e761c40f85af0fd983cab0313995d5d559767cf

https://github.com/num0005/cartographer/commit/d23e8f068c09d0c6268b72488d6d70445a1c6614

https://github.com/num0005/cartographer/commit/82c6455a8992e8d4655464901d32560b222a5f31

https://github.com/num0005/cartographer/commit/09be95aa94abb5318000c41ac08512722515e89d

-Commented out some unused features