Ranks and Networking
Reworked getPlayerRanks to utilize NetworkSession info as it populates before a game ends.

Also added basic filtering to the getPlayerRanks to not constantly spam clients with packets, and only players who join. also a force all option for when GameState reaches Lobby so that every player is updated.

Issues with this implimentation are that when a player leaves in the lobby and rejoins it will not send them their rank until a match has been played.

Added a new case to Custom packets::handle_channel_message_hook for peer_properties. This is used to leverage the ability to send a joining player thier rank as the Client requests peer properties after it has fully established itself inside a lobby. (Open to better options, but this is the best i've found).