Networking Events Statistics
Created event handling for Player Leave/Join from CustomPackets.cpp and H2MOD.cpp

Can register a callback to them which will pass a XUID parameter through it, can probably be changed in the future to have a copy of the player data or point to the NetworkSession::GetPlayerInformation

Fix an issue with rank's that if a player leaves and rejoins in lobby the rank will not appear until a match has been played.

Added a helper function to NetworkSession getPeerXUID which gets the XUID from the peer index.