Regorganizing and VIP Locking
Shifted some code around to make it more readable

Added an optional setting that will lock the server into VIP mode while the game is running. It will allow only the players who were in the lobby in the game start to rejoin when the game is going. Incase there are any network issues.

Added another function to ServerConsole, ServerConsole::SendCommand2 that for runs H2Server/H2Admin commands that is more usable than the eixsting one (SendCommand)