Wine fixes and some other changes
Changed some calls around to fix issues with StatsHandler when it is running under wine. Really don't know or understand what the issue reall was but the changes fix it. My current best guess is something with pointers or their respective data being overridden.

Removed uncessecary code in Accounts.cpp that was left over from my testing on how to recover old account files.

Commented out code to fix the visor, as I was told not to change shader templates. Leaving it there incase it's deemed okay to do in the future.