Commands, Events and Invalidation
Created an enum ServerConsole::ServerConsoleCommands that is used in conjuction with the dedi command hook.

Created a new event handler for ServerConsoleCommand that allows callbacks to be executed when a specific server command is ran.

Moved all the event registration for StatsHandler into itself and out of H2MOD.cpp.

Added a new EventHandler to statshandler for detecting when the server is told to skip the match. Which will invalidate the match and stop it from submitting it to the API.