Various
Added extra logging to Custom Packets, to check that invalid custom packets aren't being sent..

Added some now disabled code that adjusted the recovery time tick rate for weapons, it worked not as expected. Leaving there for future reversing.

force_even_teams should now be working, changed the operation so that it will sort teams constantly rather than when the countdown starts.

Made it so that team_bitflags are not applied in zombies.