Various
Added new settings
Uniform Sensitivity: makes X/Y sensitivity equal to one another
Mouse Sens: (Change) changed the setting to be a float
Controller Sens: (Change) made the setting a float
Experimental Render changes: See below

Rewrote and optimized the main_game_loop removing uneeded calls and checks to things that are not actually in the games code. Pulled the render call out of the game loop and calls it outside of it. This significantly improves the preformance of the game and allows "true" uncapped gameplay.

Refactored SetSens code out of H2Tweaks and placed them in thier respective Input namespaces

Added the ability for keyboard players to skip players on deathcam (spacebar) by leveraging a call to get_controller_input_data.