Fixes
StatsHandler was no longer initializing, moved the initializer to be called once H2MOD initializer was called.

Changed some tooltips in the advanced settings to make them less confusing or more apparent.

Removing Disable Keyboard input, will be potentially re-added once it is rewritten. The current method of doing it causes to many crashes.

Refactored and moved some code around in the ImGui Handler.

Changed the MOTD image to be loaded on start up in a seperate thread so it does not lock the games process for people with bad connections.

Created a new input blocking method so that MK and Controller cannot be used at the same time.

Uncut really grinds my gears.