More Fixes
Restored missing code from OGH2 to lock the game to the tick rate

Fixed ranks so they are not loaded during a match for servers that have them enabled, as it causes lag.

Fixed a bug with matches being invalidated when skipping to a specific match in lobby

Changed Stats json file name to be use xuid rather than server name.

Removed old code that set rank to -1

Re-enabled Projectile fix for Snipers