Fixes and Additions
Added a function to be able to send a message as the server programatically. Without sending Console command input so it also works as a service.

Removed a portion of the Dynamic Light fix so that the BSP does not have z-flighting issues while game models still cast dynamic lights.