Lots of Changes
Moved GAME_BUILD and EXECUTABLE_VERSION to defines inside of H2MOD.h

Renamed Engine:: to EngineCalls::

Created EngineHooks:: to be utilized for any hooks that are more of a global thing, part of the H2MOD.cpp cleanup

Added a function to PhysicsConstants to get the default gravity value.

Change Custom Setting Gravity to be a ratio of default gravity (Setting 1.0 = Default Gravity)

Added Gamespeed and Infinite Grenades as a setting.

Added Bluescreen detection and Player Spawn/Death Events.

Changed some Types in GunGame to use Datum instead of Int

Added a little bit more logging to StatsHandler