Blam Def Changes, Cleanup, and Custom Settings
Updated some blam def structs and enums to be more descriptive of what they actually represent.

Cleaned up some typedefs so there is less clutter.

Added a predefined settings for hill rotation to manually define which hills are used in rotation.

Added a setting for controling spawn protection time.

added a util functions for testing if a string is an integer.

re-added mapload event execution.