Lazy Blam Updates
Implemented a LazyBuffer to help with loading tag data, it expands itself to be larger while freeing old buffers.

Discovered that the instance size does not include things outside of the scope of the tag_data (i.e datarefs, opmitized tag blocks) which makes the LazyBuffer essential for loading them.