Event Fixes and first run updating
Fixed some events not firing under certain conditions.

Added some logic to force an update if no config exists, this is under the assumption that when someone doesn't have a config they are most likely a first run user. This is used to make sure they have all the content downloaded and not just the latest xlive.dll